projects/01-file-organizer/requirements.txt
watchdog==6.0.0
PyYAML==6.0.3
